TrainRuns.jl Documentation

TrainRuns.jl Documentation

Public Functions

Privat Functions

Index

Public Functions

TrainRuns.Path — Type

Path(file, type = :YAML)

Path is a datastruture for calculation context. The function Path() will create a running path for the train. Supported formats are: railtoolkit/schema (2022.05)

Example

julia> my_path = Path("file.yaml") # will generate a path from a YAML file.
Path(variables)

TrainRuns.Settings — Type

Settings(file)

Settings is a datastruture for calculation context. The function Settings() will create a set of settings for the train run calculation. file is optinal may be used to load settings in the YAML format.

Example

julia> my_settings = Settings() # will generate default settings
# massModel, stepVariable, stepSize, approxLevel, outputDetail, outputFormat
Settings(:mass_point, :distance, 20, 3, :running_time, :dataframe)

TrainRuns.Train — Type

Train(file, type = :YAML)

Train is a datastruture for calculation context. The function Train() will create a train to use in calculations. Supported formats for the YAML files are: railtoolkit/schema (2022.05)

Example

julia> my_train = Train("file.yaml") # will generate a train from a YAML file.
Train(variables)

TrainRuns.trainrun — Function

trainrun(train::Train, path::Path, settings::Settings)

Calculate the running time of a train on a path. The settings provides the choice of models for the calculation. settings can be omitted. If so, a default is used. The running time will be return in seconds.

Examples

julia> trainrun(train, path)[end,:t]
xxx.xx # in seconds

Privat Functions

TrainRuns.SupportPoint — Method

a SupportPoint is the smallest element of the driving course. One step of the step approach is between two support points

TrainRuns.acceleration — Method

acceleration(F_T, F_R, m_train, ξ_train)

Calculate the acceleration in m/s^2 with train characteristics and forces.

Arguments

F_T::Real: the tractive effort in N.

F_R::Real: the resisting forces in N.

m_train::Real: the train's mass in kg.

ξ_train::Real: the train's rotation mass factor (without unit).

Examples

julia> acceleration(94400.0, 1700.0, 88000.0, 1.08)
0.9753787878787878

TrainRuns.addAcceleratingSection! — Method

addAcceleratingSection!(drivingCourse, stateFlags, CSs, csId, settings, train)

Calculate the support points of the accelerating section and add them to drivingCourse.

See also addBreakFreeSection!, addClearingSection!, addCruisingSection!, addDiminishingSection!, addCoastingSection!, addBrakingSection!, addHalt!.

Arguments

drivingCourse::Vector{Dict}: the Vector containing dictionaries for all support points.

stateFlags::Dict: the Dictionary containing different states for drivingCourse's values.

CSs::Vector{Dict}: the Vector containing dictionaries for all characteristic sections.

csId::Integer: the number of characteristic section in the Vector CSs.

settings::Settings: the Settings object containing settings used for calculation.

train::Train: the Train object containing attributes with technical train data.

Examples

See addBreakFreeSection!.

TrainRuns.addBrakingSection! — Method

addBrakingSection!(drivingCourse, stateFlags, CSs, csId, settings, train)

Calculate the support points of the braking section and add them to drivingCourse.

See also addBreakFreeSection!, addClearingSection!, addAcceleratingSection!, addCruisingSection!, addDiminishingSection!, addCoastingSection!, addHalt!.

Arguments

drivingCourse::Vector{Dict}: the Vector containing dictionaries for all support points.

stateFlags::Dict: the Dictionary containing different states for drivingCourse's values.

CSs::Vector{Dict}: the Vector containing dictionaries for all characteristic sections.

csId::Integer: the number of characteristic section in the Vector CSs.

settings::Settings: the Settings object containing settings used for calculation.

train::Train: the Train object containing attributes with technical train data.

Examples

See addBreakFreeSection!.

TrainRuns.addBreakFreeSection! — Method

addBreakFreeSection!(drivingCourse, stateFlags, CSs, csId, settings, train)

Calculate the support points of the breakFree section and add them to drivingCourse.

See also addClearingSection!, addAcceleratingSection!, addCruisingSection!, addDiminishingSection!, addCoastingSection!, addBrakingSection!, addHalt!.

Arguments

drivingCourse::Vector{Dict}: the Vector containing dictionaries for all support points.

stateFlags::Dict: the Dictionary containing different states for drivingCourse's values.

CSs::Vector{Dict}: the Vector containing dictionaries for all characteristic sections.

csId::Integer: the number of characteristic section in the Vector CSs.

settings::Settings: the Settings object containing settings used for calculation.

train::Train: the Train object containing attributes with technical train data.

Examples

julia> addBreakFreeSection(drivingCourse_with_starting_point, stateFlags, CSs, 1, settings_default, train_longdistance)
(Dict[Dict{Symbol, Any}(:label => "", :behavior => "breakFree", :F_T => 300000, :R_wagons => 7309.09433933, :R_path => 0.0, :s => 0, :v => 0.0, :R_train => 9505.53877308, :R_traction => 2196.44443375, :a => 0.6143175668391081, :t => 0.0, :F_R => 9505.53877308), Dict{Symbol, Any}(:label => "", :behavior => "breakFree", :F_T => 300000.0, :R_wagons => 8848.180631278232, :R_path => 0.0, :s => 20, :v => 4.957086107136361, :R_train => 11471.657638238565, :R_traction => 2623.4770069603337, :a => 0.6101597548372565, :t => 8.069256643013498, :F_R => 11471.657638238565)], Dict{Symbol, Bool}(:previousSpeedLimitReached => 0, :speedLimitReached => 0, :brakingStartReached => 0, :resistingForceNegative => 0, :endOfCSReached => 0, :tractionDeficit => 0, :error => 0))

TrainRuns.addClearingSection! — Method

addClearingSection!(drivingCourse, stateFlags, CSs, csId, settings, train)

Calculate the support points of the clearing section and add them to drivingCourse.

See also addBreakFreeSection!, addCruisingSection!.

Arguments

drivingCourse::Vector{Dict}: the Vector containing dictionaries for all support points.

stateFlags::Dict: the Dictionary containing different states for drivingCourse's values.

CSs::Vector{Dict}: the Vector containing dictionaries for all characteristic sections.

csId::Integer: the number of characteristic section in the Vector CSs.

settings::Settings: the Settings object containing settings used for calculation.

train::Train: the Train object containing attributes with technical train data.

Examples

See addBreakFreeSection!.

TrainRuns.addCoastingSection! — Method

addCoastingSection!(drivingCourse, stateFlags, CSs, csId, settings, train)

Calculate the support points of the coasting section and add them to drivingCourse.

See also addBreakFreeSection!, addClearingSection!, addAcceleratingSection!, addCruisingSection!, addDiminishingSection!, addBrakingSection!, addHalt!.

Arguments

drivingCourse::Vector{Dict}: the Vector containing dictionaries for all support points.

stateFlags::Dict: the Dictionary containing different states for drivingCourse's values.

CSs::Vector{Dict}: the Vector containing dictionaries for all characteristic sections.

csId::Integer: the number of characteristic section in the Vector CSs.

settings::Settings: the Settings object containing settings used for calculation.

train::Train: the Train object containing attributes with technical train data.

Examples

See addBreakFreeSection!.

TrainRuns.addCruisingSection! — Method

addCruisingSection!(drivingCourse, stateFlags, CSs, csId, settings, train, cruisingType, s_cruising)

Calculate the support points of the cruising section and add them to drivingCourse.

See also addBreakFreeSection!, addClearingSection!, addAcceleratingSection!, addDiminishingSection!, addCoastingSection!, addBrakingSection!, addHalt!.

Arguments

drivingCourse::Vector{Dict}: the Vector containing dictionaries for all support points.

stateFlags::Dict: the Dictionary containing different states for drivingCourse's values.

CSs::Vector{Dict}: the Vector containing dictionaries for all characteristic sections.

csId::Integer: the number of characteristic section in the Vector CSs.

settings::Settings: the Settings object containing settings used for calculation.

train::Train: the Train object containing attributes with technical train data.

cruisingType::String: the type of crusing respectively if applicable the reason why the train has to cruise

s_cruising::Real: the planned maximum crusing distance

Examples

See addBreakFreeSection!.

TrainRuns.addDiminishingSection! — Method

addDiminishingSection!(drivingCourse, stateFlags, CSs, csId, settings, train)

Calculate the support points of the diminishing section (when using maximum tractive effort and still getting slower) and add them to drivingCourse.

See also addBreakFreeSection!, addClearingSection!, addAcceleratingSection!, addCruisingSection!, addCoastingSection!, addBrakingSection!, addHalt!.

Arguments

drivingCourse::Vector{Dict}: the Vector containing dictionaries for all support points.

stateFlags::Dict: the Dictionary containing different states for drivingCourse's values.

CSs::Vector{Dict}: the Vector containing dictionaries for all characteristic sections.

csId::Integer: the number of characteristic section in the Vector CSs.

settings::Settings: the Settings object containing settings used for calculation.

train::Train: the Train object containing attributes with technical train data.

Examples

See addBreakFreeSection!.

TrainRuns.addHalt! — Method

addHalt!(drivingCourse, stateFlags, CSs, csId, settings, train)

Calculate the support points of the halt section and add them to drivingCourse.

See also addBreakFreeSection!, addClearingSection!, addAcceleratingSection!, addCruisingSection!, addDiminishingSection!, addCoastingSection!, addBrakingSection!.

Arguments

drivingCourse::Vector{Dict}: the Vector containing dictionaries for all support points.

stateFlags::Dict: the Dictionary containing different states for drivingCourse's values.

CSs::Vector{Dict}: the Vector containing dictionaries for all characteristic sections.

csId::Integer: the number of characteristic section in the Vector CSs.

settings::Settings: the Settings object containing settings used for calculation.

train::Train: the Train object containing attributes with technical train data.

Examples

See addBreakFreeSection!.

TrainRuns.brakingAcceleration — Method

brakingAcceleration(v_start, v_end, s_braking)

Calculate the acceleration in m/s^2 to decelerate from v_start to v_end in m/s on s_braking in m.

See also brakingDistance, brakingStartVelocity.

Examples

julia> brakingAcceleration(25.0, 15.0, 500)
-0.4

TrainRuns.brakingDistance — Method

brakingDistance(v_start, v_end, a_braking, approxLevel)

Calculate the braking distance in m with velocities in m/s and acceleration in m/s^2.

See also brakingAcceleration, brakingStartVelocity.

Arguments

v_start::Real: the velocity at the start of braking in m/s.

v_end::Real: the target velocity at the end of braking in m/s.

a_braking::Real: the constant braking acceleration in m/s^2.

approxLevel::Integer: the last position behind the decimal point that is not rounded

Examples

julia> brakingDistance(25.0, 15.0, -0.4253, 3)
470.2563

TrainRuns.brakingStartVelocity — Method

brakingStartVelocity(v_end, a_braking, s_braking, approxLevel)

Calculate the maximum velocity in m/s where the train can start to brake to reach v_end.

See also brakingAcceleration, brakingDistance.

Arguments

v_end::Real: the target velocity at the end of braking in m/s.

a_braking::Real: the constant braking acceleration in m/s^2.

s_braking::Real: the braking distance in m.

approxLevel::Integer: the last position behind the decimal point that is not rounded

Examples

julia> brakingStartVelocity(15.0, -0.4253, 500, 3)
25.4656

TrainRuns.calculateForces! — Method

calculate and return tractive and resisting forces for a support point

TrainRuns.calculatePathResistance — Method

calculate and return the path resistance dependend on the trains position and mass model

TrainRuns.calculateTractiveEffort — Method

calculateTractiveEffort(v, tractiveEffortVelocityPairs)

Calculate the trains tractive effort with the tractiveEffortVelocityPairs dependend on the velocity v.

...

Arguments

v::AbstractFloat: the current velocity in m/s.

tractiveEffortVelocityPairs::Array{}: the trains pairs for velocity in m/s and tractive effort in N as one array containing an array for each pair.

...

Examples

julia> calculateTractiveEffort(20.0, [(0.0, 180000), (20.0, 100000), (40.0, 60000), (60.0, 40000), (80.0, 30000)])
100000

julia> calculateTractiveEffort(30.0, [(0.0, 180000), (20.0, 100000), (40.0, 60000), (60.0, 40000), (80.0, 30000)])
80000

TrainRuns.createDataFrame — Method

createDataFrame(output_vector, outputDetail, approxLevel)

Create a DataFrame from output_vector with outputDetail and approxLevel.

See also createOutput.

Arguments

output_vector::Vector{Dict}: the Vector containing all data to be outputted.

outputDetail::Symbol: the detail level the DataFrame is created for.

approxLevel::Int: the number of digits for rounding each Number in the DataFrame.

Examples

julia> createDataFrame(vector_pointsOfInterest, detail_data_points, approxLevel_default)
5×11 DataFrame
 Row │ label             driving_mode  s        v       t        a       F_T        F_R      R_path   R_traction  R_wagons
     │ String            String        Real     Real    Real     Real    Real       Real     Real     Real        Real
─────┼─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ view_point_1      accelerating   850.0   28.707   54.049   0.331  1.93049e5  36602.1     0.0      9088.56   27513.6
   2 │ distant_signal_1  accelerating  1000.0   30.325   59.129   0.294  1.82746e5  43604.7  4344.35     9795.13   29465.2
   3 │ main_signal_1     accelerating  2000.0   37.356   88.468   0.185  1.48352e5  60899.4  8688.69    13259.1    38951.5
   4 │ main_signal_3     braking       9000.0   27.386  258.578  -0.375  0.0        34522.1     0.0      8537.05   25985.0
   5 │ clearing_point_1  braking       9203.37  24.443  266.426  -0.375  0.0        30176.2     0.0      7389.44   22786.8

TrainRuns.createOutput — Method

createOutput(settings, drivingCourse, pointsOfInterest)

Create output information depending on settings, drivingCourse and pointsOfInterest.

See also createOutput.

Arguments

settings::Settings: the Settings object containing settings for output format and detail.

drivingCourse::Vector{Dict}: the Vector containing dictionaries for all support points.

pointsOfInterest::Vector{NamedTuple}: the Vector containing tuples for the paths' points of interest.

Examples

julia> createOutput(settings_poi, drivingCourse_longdistance, pointsOfInterest_pathWithSlope)
5×11 DataFrame
 Row │ label             driving_mode  s        v       t        a       F_T        F_R      R_path   R_traction  R_wagons
     │ String            String        Real     Real    Real     Real    Real       Real     Real     Real        Real
─────┼─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ view_point_1      accelerating   850.0   28.707   54.049   0.331  1.93049e5  36602.1     0.0      9088.56   27513.6
   2 │ distant_signal_1  accelerating  1000.0   30.325   59.129   0.294  1.82746e5  43604.7  4344.35     9795.13   29465.2
   3 │ main_signal_1     accelerating  2000.0   37.356   88.468   0.185  1.48352e5  60899.4  8688.69    13259.1    38951.5
   4 │ main_signal_3     braking       9000.0   27.386  258.578  -0.375  0.0        34522.1     0.0      8537.05   25985.0
   5 │ clearing_point_1  braking       9203.37  24.443  266.426  -0.375  0.0        30176.2     0.0      7389.44   22786.8

TrainRuns.forceFromCoefficient — Method

forceFromCoefficient(f, m)

Calculate the force in N with the coefficient f in ‰ and the vehicle's mass m in kg.

Examples

julia> forceFromCoefficient(1.5, 68000.0)
1000.2783

TrainRuns.freightWagonsResistance — Method

freightWagonsResistance(v, train)

Calculate the vehicle resistance in N for the freight wagons of the train dependend on the velocity v.

Arguments

v::AbstractFloat: the current velocity in m/s.

train::Train: the struct defined in types.jl containing attributes with technical data.

Examples

julia> freightWagonsResistance(15.0, freight_train)
20900.732702639998

TrainRuns.getLowestSpeedLimit — Method

if the rear of the train is still located in a former characteristic section it has to be checked if its speed limit can be kept

TrainRuns.getNextPointOfInterest — Method

TODO

TrainRuns.moveAStep — Method

TODO

TrainRuns.passengerWagonsResistance — Method

passengerWagonsResistance(v, train)

Calculate the vehicle resistance in N for the passenger wagons of the train dependend on the velocity v.

Arguments

v::AbstractFloat: the current velocity in m/s.

train::Train: the struct defined in types.jl containing attributes with technical data.

Examples

julia> passengerWagonsResistance(15.0, longdistance_passenger_train)
14461.2708244928

TrainRuns.recalculateLastBrakingPoint! — Method

recalculateLastBrakingPoint!(drivingCourse, s_target, v_target)

Realculate the last step of the drivingCourse depending on new position and velocity.

Arguments

drivingCourse::Vector{Dict}: the Vector containing dictionaries for all support points.

s_target::Real: the new positon for the last support point in m.

v_target::Real: the new velocity for the last support point in m/s.

Examples

julia> recalculateLastBrakingPoint(drivingCourse_with_two_points_near_5000m, 5000.0, 23.333333333333332)
Dict{Symbol, Any}[Dict(:label => "", :behavior => "braking", :F_T => 0.0, :R_wagons => 21655.0, :R_path => 0.0, :s => 4999.999980000006, :v => 23.333333441372545, :R_train => 28640.8, :R_traction => 6985.8, :a => -0.126, :t => 235.16501930395856, :F_R => 28640.8), Dict(:label => "", :behavior => "braking", :F_T => 0.0, :R_wagons => 0.0, :R_path => 0.0, :s => 5000.0, :v => 23.333333333333332, :R_train => 0.0, :R_traction => 0.0, :a => 0.0, :t => 235.16502016110115, :F_R => 0.0)]

TrainRuns.secureBrakingBehavior! — Method

secureBrakingBehavior!(CSs, a_braking, approxLevel)

Calculate the velocities between the characterisitc sections to secure braking behavior.

Arguments

CSs::Vector{Dict}: the Vector containing dictionaries for all characteristic sections.

a_braking::Real: the constant braking acceleration in m/s^2.

approxLevel::Integer: the last position behind the decimal point that is not rounded.

Examples

julia> secureBrakingBehavior!(CSs, -0.375, 3)
Dict[Dict{Symbol, Any}(:s_entry => 0.0, :v_exit => 19.3649, :v_limit => 25.0, :r_path => 0.0, :pointsOfInterest => NamedTuple[(s = 9500.0, label = "")], :s_exit => 9500.0), Dict{Symbol, Any}(:s_entry => 0.0, :v_exit => 0.0, :v_limit => 25.0, :r_path => 1.0, :pointsOfInterest => NamedTuple[(s = 10000.0, label = "")], :s_exit => 10000.0)]

TrainRuns.tractionUnitResistance — Method

tractionUnitResistance(v, train)

Calculate the vehicle resistance in N for the traction unit of the train dependend on the velocity v.

Arguments

v::AbstractFloat: the current velocity in m/s.

train::Train: the struct defined in types.jl containing attributes with technical data.

Examples

julia> tractionUnitResistance(15.0, freight_train)
5461.127252

TrainRuns.Δs_with_Δt — Method

Δs_with_Δt(Δt, a_prev, v_prev)

Calculate the distance step in m with the time step in s, acceleration in m/s^2 and velocity in m/s.

See also Δs_with_Δv, Δt_with_Δs, Δt_with_Δv, Δt_with_constant_v, Δv_with_Δs, Δv_with_Δt.

Examples

julia> Δs_with_Δt(3.0, 0.5, 25.0)
77.25

TrainRuns.Δs_with_Δv — Method

Δs_with_Δv(Δv, a_prev, v_prev)

Calculate the distance step in m with the velocity step in m/s, acceleration in m/s^2 and velocity in m/s.

See also Δs_with_Δt, Δt_with_Δs, Δt_with_Δv, Δt_with_constant_v, Δv_with_Δs, Δv_with_Δt.

Examples

julia> Δs_with_Δv(1.0, 0.5, 25.0)
51.0

TrainRuns.Δt_with_constant_v — Method

Δt_with_constant_v(Δs, v)

Calculate the time step in s with the distance step in m and constant velocity in m/s.

See also Δs_with_Δt, Δs_with_Δv, Δt_with_Δs, Δt_with_Δv, Δv_with_Δs, Δv_with_Δt.

Examples

julia> Δt_with_constant_v(10.0, 25.0)
0.4

TrainRuns.Δt_with_Δs — Method

Δt_with_Δs(Δs, a_prev, v_prev)

Calculate the time step in s with the distance step in m, acceleration in m/s^2 and velocity in m/s.

See also Δs_with_Δt, Δs_with_Δv, Δt_with_Δv, Δt_with_constant_v, Δv_with_Δs, Δv_with_Δt.

Examples

julia> Δt_with_Δs(10.0, 0.5, 25.0)
0.39841267341660824

TrainRuns.Δt_with_Δv — Method

Δt_with_Δv(Δv, a_prev)

Calculate the time step in s with the velocity step in m/s and the acceleration in m/s^2.

See also Δs_with_Δt, Δs_with_Δv, Δt_with_Δs, Δt_with_constant_v, Δv_with_Δs, Δv_with_Δt.

Examples

julia> Δt_with_Δv(1.0, 0.5)
2.0

TrainRuns.Δv_with_Δs — Method

Δv_with_Δs(Δs, a_prev, v_prev)

Calculate the velocity step in m/s with the distance step in m, acceleration in m/s^2 and velocity in m/s.

See also Δs_with_Δt, Δs_with_Δv, Δt_with_Δs, Δt_with_Δv, Δt_with_constant_v, Δv_with_Δt.

Examples

julia> Δv_with_Δs(10.0, 0.5, 25.0)
0.19920633670830412

TrainRuns.Δv_with_Δt — Method

Δv_with_Δt(Δt, a_prev, v_prev)

Calculate the velocity step in m/s with the time step in s and acceleration in m/s^2.

See also Δs_with_Δt, Δs_with_Δv, Δt_with_Δs, Δt_with_Δv, Δt_with_constant_v, Δv_with_Δs.

Examples

julia> Δv_with_Δt(3.0, 0.5)
1.5

Index

TrainRuns.SupportPoint

TrainRuns.acceleration

TrainRuns.addAcceleratingSection!

TrainRuns.addBrakingSection!

TrainRuns.addBreakFreeSection!

TrainRuns.addClearingSection!

TrainRuns.addCoastingSection!

TrainRuns.addCruisingSection!

TrainRuns.addDiminishingSection!

TrainRuns.addHalt!

TrainRuns.brakingAcceleration

TrainRuns.brakingDistance

TrainRuns.brakingStartVelocity

TrainRuns.calculateForces!

TrainRuns.calculatePathResistance

TrainRuns.calculateTractiveEffort

TrainRuns.createDataFrame

TrainRuns.createOutput

TrainRuns.forceFromCoefficient

TrainRuns.freightWagonsResistance

TrainRuns.getLowestSpeedLimit

TrainRuns.getNextPointOfInterest

TrainRuns.moveAStep

TrainRuns.passengerWagonsResistance

TrainRuns.recalculateLastBrakingPoint!

TrainRuns.secureBrakingBehavior!

TrainRuns.tractionUnitResistance

TrainRuns.trainrun

TrainRuns.Δs_with_Δt

TrainRuns.Δs_with_Δv

TrainRuns.Δt_with_constant_v

TrainRuns.Δt_with_Δs

TrainRuns.Δt_with_Δv

TrainRuns.Δv_with_Δs

TrainRuns.Δv_with_Δt

TrainRuns.Path

TrainRuns.Settings

TrainRuns.Train